move event handling to subclass.
