+++ electron.ts
@@ -1,13 +1,13 @@
 'use strict';
-import {app, dialog, ipcMain} from 'electron';
+import {app, ipcMain} from 'electron';
 var _log, log         = () => { _log = _log || require('electron-log'); return _log; };
-var _path, path       = () => { _path = _path || require('path'); return _path; };
 var _monitor, monitor = () => { _monitor = _monitor || require('electron-performance-monitor'); return _monitor; };
 
 import Menu from './electron-menu';
 import Pane from './electron-window';
 import Launch from './electron-launch';
 import AppConfig from './electron-appcfg';
+import EventHandler from './electron-event';
 
 // env
 const DEBUG = process.env.DEBUG != null;
@@ -53,8 +53,8 @@ const MYukkuriVoice = function(): void {
 };
 const myApp = new MYukkuriVoice() as yubo.IMYukkuriVoice;
 // set prototype
-// MYukkuriVoice.prototype.fnc = (Pane, Menu, Launch, AppConfig).prototype.fnc
-[Pane, Menu, Launch, AppConfig].forEach(baseCtor => {
+// MYukkuriVoice.prototype.fnc = (Pane, Menu, Launch, AppConfig, EventHandler).prototype.fnc
+[Pane, Menu, Launch, AppConfig, EventHandler].forEach(baseCtor => {
   Object.getOwnPropertyNames(baseCtor.prototype).forEach(name => {
     if (name != 'constructor') {
       Object.defineProperty(MYukkuriVoice.prototype, name, Object.getOwnPropertyDescriptor(baseCtor.prototype, name));
@@ -98,6 +98,8 @@ app.on('ready', () => {
     myApp.showMainWindow();
     myApp.initAppMenu();
     myApp.initDockMenu();
+    // register event handler
+    myApp.acceptEvents();
     if (MONITOR) { log().warn(monitor().format('electron', 'ready done')); }
   });
 });
@@ -110,120 +112,3 @@ if (MONITOR) {
   });
 }
 
-// show window event
-ipcMain.on('showHelpWindow', (event: Electron.Event, message: string) => {
-  myApp.showHelpWindow();
-});
-ipcMain.on('showHelpSearchDialog', (event: Electron.Event, message: string) => {
-  myApp.showHelpSearchDialog();
-});
-ipcMain.on('showSystemWindow', (event: Electron.Event, message: string) => {
-  myApp.showSystemWindow();
-});
-ipcMain.on('showDictWindow', (event: Electron.Event, message: string) => {
-  myApp.showDictWindow();
-});
-ipcMain.on('showSpecWindow', (event: Electron.Event, message: string) => {
-  myApp.showSpecWindow();
-});
-
-// showSaveDialog
-ipcMain.on('showSaveDialog', (event: Electron.Event, message: string) => {
-  const options = {
-    title: 'wav save dialog',
-    filters: [
-      {name: 'Wav File', extensions: ['wav']},
-    ],
-  };
-  const r = dialog.showSaveDialog(myApp.mainWindow, options);
-  event.sender.send('showSaveDialog', r);
-});
-
-// showDirDialog
-ipcMain.on('showDirDialog', (event: Electron.Event, defaultPath: string) => {
-  const options = {
-    title: 'select wav save directory',
-    properties: ['openDirectory' as 'openDirectory', 'createDirectory' as 'createDirectory'],
-    defaultPath: defaultPath,
-  };
-  const r = dialog.showOpenDialog(myApp.mainWindow, options);
-  event.sender.send('showDirDialog', r);
-});
-
-// drag out wav file
-ipcMain.on('ondragstartwav', (event: Electron.Event, filePath: string) => {
-  const imgPath = path().join(__dirname, '/images/ic_music_video_black_24dp_1x.png');
-  event.sender.startDrag({
-    file: filePath,
-    icon: imgPath,
-  });
-});
-
-// updateAppConfig
-ipcMain.on('updateAppConfig', (event: Electron.Event, options: yubo.AppCfg) => {
-  myApp.updateAppConfig(options);
-  const dialogOptions = {
-    type: 'info',
-    title: 'application config updated.',
-    message: '環境設定を更新しました。アプリケーションを更新します。',
-    buttons: ['OK'],
-    defaultId: 0,
-  };
-  const r = dialog.showMessageBox(myApp.systemWindow, dialogOptions);
-  event.sender.send('updateAppConfig', r);
-  myApp.mainWindow.setSize(myApp.appCfg.mainWindow.width, myApp.appCfg.mainWindow.height);
-  myApp.mainWindow.webContents.reload();
-  if (myApp.systemWindow) { myApp.systemWindow.webContents.reload(); }
-});
-
-// resetAppConfig
-ipcMain.on('resetAppConfig', (event: Electron.Event, message: string) => {
-  myApp.resetAppConfig();
-  const dialogOptions = {
-    type: 'info',
-    title: 'application config initialized.',
-    message: '環境設定を初期化しました。アプリケーションを更新します。',
-    buttons: ['OK'],
-    defaultId: 0,
-  };
-  const r = dialog.showMessageBox(myApp.systemWindow, dialogOptions);
-  event.sender.send('resetAppConfig', r);
-  myApp.mainWindow.setSize(myApp.appCfg.mainWindow.width, myApp.appCfg.mainWindow.height);
-  myApp.mainWindow.webContents.reload();
-  if (myApp.systemWindow) { myApp.systemWindow.webContents.reload(); }
-});
-
-// resetAppConfigOnMain
-function resetAppConfigOnMain(): void {
-  myApp.resetAppConfig();
-  const dialogOptions = {
-    type: 'info',
-    title: 'application config initialized.',
-    message: '環境設定を初期化しました。アプリケーションを更新します。',
-    buttons: ['OK'],
-    defaultId: 0,
-  };
-  const r = dialog.showMessageBox(myApp.mainWindow, dialogOptions);
-  myApp.mainWindow.setSize(myApp.appCfg.mainWindow.width, myApp.appCfg.mainWindow.height);
-  myApp.mainWindow.webContents.reload();
-  if (myApp.systemWindow) { myApp.systemWindow.webContents.reload(); }
-}
-MYukkuriVoice.prototype.resetAppConfigOnMain = resetAppConfigOnMain;
-
-ipcMain.on('reloadMainWindow', (event: Electron.Event, message: string) => {
-  myApp.mainWindow.webContents.reload();
-  event.sender.send('reloadMainWindow', message);
-});
-
-// switchAlwaysOnTop
-function switchAlwaysOnTop(): void {
-  const newflg = !myApp.mainWindow.isAlwaysOnTop();
-  myApp.mainWindow.setAlwaysOnTop(newflg);
-  myApp.mainWindow.webContents.send('switchAlwaysOnTop', newflg);
-}
-ipcMain.on('switchAlwaysOnTop', (event: Electron.Event, message: string) => {
-  const newflg = !myApp.mainWindow.isAlwaysOnTop();
-  myApp.mainWindow.setAlwaysOnTop(newflg);
-  event.sender.send('switchAlwaysOnTop', newflg);
-});
-MYukkuriVoice.prototype.switchAlwaysOnTop = switchAlwaysOnTop;
